+++ src/main/java/com/ssb/ajax/action/ajaxController.java
@@ -60,22 +60,6 @@ public class ajaxController extends HttpServlet {
 			//정보처리
 			int result = dao.deleteCart(cart_id,member_id);
 			json = gson.toJson(result);
-		} else if (command.equals("/insertWishlist.aj")) {
-			// 정보저장
-			String item_id = request.getParameter("item_id");
-			String member_id = (String)request.getSession().getAttribute("member_id");
-			System.out.println("item_id : " + item_id);
-			System.out.println("member_id : " + member_id);
-			//정보처리
-			String result = dao.insertWishlist(item_id,member_id);
-			json = gson.toJson(result);
-		} else if (command.equals("/getWishlist.aj")) {
-			// 정보저장
-			int member_id = Integer.parseInt((String)request.getSession().getAttribute("member_id"));
-			System.out.println("member_id : " + member_id);
-			//정보처리
-			ArrayList<Integer> result = dao.getWishlist(member_id);
-			json = gson.toJson(result);
 		}
 		
 		/*********************** 2. 가상주소 매핑 끝 **************************/
