+++ src/main/java/com/ssb/ajax/db/ajaxDAO.java
@@ -74,64 +74,5 @@ public class ajaxDAO extends DAO {
 		return result;
 	}
 
-	public String insertWishlist(String item_id, String member_id) {
-		String result = null;
-		try {
-			con = getCon();
-			sql = "SELECT * FROM wishlist WHERE item_id IN("+ item_id +") AND member_id = ?";
-			pstmt = con.prepareStatement(sql);
-			pstmt.setString(1, member_id);
-			System.out.println(pstmt);
-			rs = pstmt.executeQuery();
-			if (rs.next()) {
-				while (rs.next()) {
-					sql = "DELETE FROM wishlist WHERE wishlist_id = ?";
-					pstmt = con.prepareStatement(sql);
-					pstmt.setInt(1, rs.getInt("wishlist_id"));
-					System.out.println(pstmt);
-					if(pstmt.executeUpdate() == 1) {
-						result = "deleted";
-					}else{
-						result = "deletedFail";
-					}
-				}
-			}else {
-				sql = "INSERT INTO wishlist VALUES(DEFAULT,?,?)";
-				pstmt = con.prepareStatement(sql);
-				pstmt.setString(1, item_id);
-				pstmt.setString(2, member_id);
-				System.out.println(pstmt);
-				if(pstmt.executeUpdate() == 1) {
-					result = "inserted";
-				}else{
-					result = "insertedFail";
-				}
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			CloseDB();
-		}
-		return result;
-	}
-
-	public ArrayList<Integer> getWishlist(int member_id) {
-		ArrayList<Integer> item_idArray = new ArrayList<Integer>();
-		try {
-			con = getCon();
-			sql = "SELECT item_id FROM wishlist WHERE member_id = ? ORDER BY wishlist_id";
-			pstmt = con.prepareStatement(sql);
-			pstmt.setInt(1, member_id);
-			rs = pstmt.executeQuery();
-			while (rs.next()) {
-				item_idArray.add(rs.getInt("item_id"));
-				System.out.println("위시리스트 추가");
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			CloseDB();
-		}
-		return item_idArray;
-	}
+	
 }
